Reflect moved functions for imports
