+++ src/toil/utils/toilLaunchCluster.py
@@ -20,7 +20,7 @@ from typing import Dict, List, Tuple, Union
 from toil import applianceSelf
 from toil.common import parser_with_common_options
 from toil.lib.aws import build_tag_dict_from_env
-from toil.options import opt_strtobool
+from toil.lib.conversions import opt_strtobool
 from toil.provisioners import (check_valid_node_types,
                                cluster_factory,
                                parse_node_types)
